fix: 获取服务端ai配置放到 useAlGlobalConfig里调用
