+++ app/renderer/src/main/src/pages/ai-re-act/hooks/useAIGlobalConfig.ts
@@ -1,9 +1,17 @@
 import { useCreation, useMemoizedFn } from 'ahooks'
 import { useEffect, useMemo, useRef, useState } from 'react'
-import { AIGlobalConfig, grpcGetAIGlobalConfig, grpcSetAIGlobalConfig } from '@/pages/ai-agent/aiModelList/utils'
+import {
+  AIGlobalConfig,
+  ServerAIGlobalConfig,
+  grpcGetAIGlobalConfig,
+  grpcSetAIGlobalConfig,
+} from '@/pages/ai-agent/aiModelList/utils'
+import { apiSystemConfig } from '@/components/layout/utils'
 import { useAIGlobalConfigStore } from '@/store/aiGlobalConfig'
+import { AIModelTypeEnum } from '@/pages/ai-agent/defaultConstant'
 import { shallow } from 'zustand/shallow'
 import { cloneDeep } from 'lodash'
+import { getFileNameByModelType } from '@/pages/ai-agent/aiModelList/aiModelForm/AIModelForm'
 
 interface UseAIGlobalConfigData {
   aiGlobalConfig: AIGlobalConfig
@@ -20,6 +28,8 @@ interface UseAIGlobalConfigEvents {
   setAIGlobalConfig: (data: Partial<AIGlobalConfig>) => Promise<void>
   /** 获取最新的值,不会设置全局变量中的值,会设置ref */
   getLastAIGlobalConfig: () => Promise<AIGlobalConfig>
+  /** 企业版登录后获取服务端 AI 配置并合并到本地配置 */
+  getAIGlobalConfigAfterLogin: () => Promise<void>
   /** 设置到全局变量中 */
   setConfigStore: (data: AIGlobalConfig) => void
 }
@@ -110,6 +120,37 @@ function useAIGlobalConfig(params) {
     aiGlobalConfigRef.current = cloneDeep(data)
   })
 
+  /** 企业版登录后获取服务端 AI 配置并合并到本地配置 */
+  const getAIGlobalConfigAfterLogin = useMemoizedFn(async () => {
+    try {
+      const currentConfig = await getLastAIGlobalConfig()
+      const nextConfig: AIGlobalConfig = cloneDeep(currentConfig)
+
+      const data = (await apiSystemConfig()).data || []
+      const aiConfig = data.find((item) => item.configName === 'AI')
+      if (!aiConfig?.content) return
+
+      let serverConfig: ServerAIGlobalConfig = {}
+      try {
+        serverConfig = JSON.parse(aiConfig.content)
+      } catch (error) {}
+
+      Object.values(AIModelTypeEnum).forEach((type) => {
+        const key = getFileNameByModelType(type)
+        if (!key) return
+        const localModels = (currentConfig[key] || []).filter((model) => !model.IsOnline)
+        const serverModels = (serverConfig[key] || []).map((model) => ({
+          ...model,
+          IsOnline: true,
+        }))
+
+        nextConfig[key] = [...localModels, ...serverModels]
+      })
+
+      setAIGlobalConfig(nextConfig)
+    } catch (err) {}
+  })
+
   const data: UseAIGlobalConfigData = useMemo(() => {
     return {
       aiGlobalConfig,
@@ -125,6 +166,7 @@ function useAIGlobalConfig(params) {
       onRefresh: (isShowLoading) => getAIGlobalConfig(isShowLoading),
       setAIGlobalConfig: (res) => setAIGlobalConfig(res),
       getLastAIGlobalConfig: () => getLastAIGlobalConfig(),
+      getAIGlobalConfigAfterLogin: () => getAIGlobalConfigAfterLogin(),
       setConfigStore: (res) => setConfigStore(res),
     }
   }, [])
